Fixed actions and added rewards
Added mixed actions (now the flight can choose both to change accelleration and angle at the same time)
Changed architecture: before we had [512, 1024, 512 ] which is way too big for an input size of 9.
Changed activation: tanh instead of Relu since we also have negative values.
Using 2 separate architectures for value function and actions
Increased a bit the entropy coefficient to allow exploration (before it was 0.0)
Updated the drift to match the one in the competition github page
Added a penalty if an agent tries to accellerate or decellerate outside the min and max allowable speed
Added a penalty if the agent moves away from the optimal trajectory (straight line from start position to target)
Added a small penalty when agent change angle (It should reduce the number of times it change angle, expecially if there is no purpose in doing it)