Removed RayWrapper, added optional observation, checked parameters
removed the RayWrapper now things are handled inside the FlightEnv itself. Now some observations are optional, for example we ca use or not the FoV and the drift. Tested different hyperparameters and found out that entropy has a big impact on the training especially if the rewards are small. Also the kl coefficient is set to zero, since we already have the ppo clip